+++ src/ui.rs
@@ -19,14 +19,10 @@ use std::sync::{Mutex, MutexGuard};
 use std::{fmt, io};
 
 use atty::Stream;
-use jujutsu_lib::commit::Commit;
-use jujutsu_lib::op_store::WorkspaceId;
-use jujutsu_lib::repo::RepoRef;
 use jujutsu_lib::repo_path::{RepoPath, RepoPathComponent, RepoPathJoin};
 use jujutsu_lib::settings::UserSettings;
 
 use crate::formatter::{ColorFormatter, Formatter, PlainTextFormatter};
-use crate::templater::TemplateFormatter;
 
 pub struct Ui<'a> {
     cwd: PathBuf,
@@ -187,31 +183,6 @@ impl<'stdout> Ui<'stdout> {
         formatter.add_label(String::from("error"))?;
         formatter.write_str(text)?;
         formatter.remove_label()?;
-        Ok(())
-    }
-
-    pub fn write_commit_summary(
-        &mut self,
-        repo: RepoRef,
-        workspace_id: &WorkspaceId,
-        commit: &Commit,
-    ) -> io::Result<()> {
-        let template_string = self
-            .settings
-            .config()
-            .get_string("template.commit_summary")
-            .unwrap_or_else(|_| {
-                if self.settings.enable_open_commits() {
-                    String::from(r#"label(if(open, "open"), commit_id.short() " " description.first_line())"#)
-                } else {
-                    String::from(r#"commit_id.short() " " description.first_line()"#)
-                }
-            });
-        let template =
-            crate::template_parser::parse_commit_template(repo, workspace_id, &template_string);
-        let mut formatter = self.stdout_formatter();
-        let mut template_writer = TemplateFormatter::new(template, formatter.as_mut());
-        template_writer.format(commit)?;
         Ok(())
     }
